move ziren keccak state definition to the main ziren file
